requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.3.0
pandas==2.2.2
openai==1.30.5
weasyprint==62.3
python-dotenv==1.0.1